# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Streaming App API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/public/superstar-stories": {
            "get": {
                "tags": [
                    "Public API"
                ],
                "summary": "Get all superstar stories",
                "description": "Get paginated list of all active superstar stories with superstar info",
                "operationId": "c05c8441c0f989ad844d840b7a487f5a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "postedby_userid": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "file_type": {
                                                        "type": "string",
                                                        "example": "image"
                                                    },
                                                    "url_path": {
                                                        "type": "string",
                                                        "example": "stories/image123.jpg"
                                                    },
                                                    "timestap": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "superstar": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "display_name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "superstar123"
                                                            },
                                                            "profile_image": {
                                                                "type": "string",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 30
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/public/superstar-posts": {
            "get": {
                "tags": [
                    "Public API"
                ],
                "summary": "Get all superstar posts",
                "description": "Get paginated list of all superstar posts with superstar info",
                "operationId": "b0d010c6aabd3a809040053b90635d29",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "superstar_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "My New Post"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "This is my post description"
                                                    },
                                                    "file_type": {
                                                        "type": "string",
                                                        "example": "image"
                                                    },
                                                    "url_path": {
                                                        "type": "string",
                                                        "example": "posts/image123.jpg"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "superstar": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "display_name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "superstar123"
                                                            },
                                                            "profile_image": {
                                                                "type": "string",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 30
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}